Move ReviewsConfiguration to lib folder. It is not really a model Make tests pass
