+++ lib/solidus_reviews/config.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'solidus_reviews/reviews_configuration'
+
 module Spree
   module Reviews
     Config = Spree::ReviewsConfiguration.new
